LayoutTests/platform/ios-sim/fast/events/keydown-numpad-keys-expected.txt


This tests keyboard events with KeyLocationCode argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 33
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 34
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 36
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 35
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 37
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 39
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 38
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 40
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 45
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 46
PASS lastKeyboardEvent.keyLocation is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 33
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 34
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 36
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 35
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 37
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 39
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 38
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 40
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 45
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 46
FAIL lastKeyboardEvent.keyLocation should be 3. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
